changed threads to work on independent subtrees
this crashes and needs to be fixed.

at first, a serial discovery of the paths under a set of paths (or a
single root) is required to make enough subtrees on which to run a given
number of threads.

because the symlink resolution is done post-hoc, nominally the threads
which are started on nonintersecting subtrees are independent. there may
be shared references between the File objects constructed, but those are
never called.

because the processes are independent, no threadsafe collective object (
a Queue) needs to be used. they are passed the input data.