fix for previous commit on symlinks
commit dcf6399c55b41af24ca2c7d005ac4a75bbfac4c9m changed the way the
class dictionary stored the files from inode number to string to find
the parent based on the file path. this messed up how symlinks were
handled in the python-only implementation. the python-only
implementation has been returned to its normal state. the C program has
been changed to include the inode both of the file and what it points
to, in the case it is a symlink.